Add files via upload
Fixed: 
Required: The reset button is working properly but it is not interactive- there's no change in size or color upon hovering. There's no indication which bookmark is active or inactive. Please update this by changing its colors when clicked or not.
For more information, visit Controls & Navigation

Fixed: 
Required: I reiterate that the interactivity needs to be set. It's a part of the specification which reads "...core formatting capabilities of buttons"